feat: PSJekyll Dockerfile and Container.init.ps1 ( Fixes #11, Fixes #12 )
Exposing more environment variables and refactoring away ruby base image

+++ Container.init.ps1
@@ -23,7 +23,6 @@
     (this does nothing, but most likely will be used in the future)
 #>
 using namespace 'mcr.microsoft.com/powershell AS powerShell'
-using namespace 'ruby:3.3.5-slim-bullseye AS jekyllrb'
 
 param(
 # The name of the module to be installed.
@@ -37,29 +36,45 @@ param(
     }
 ),
 # The packages to be installed.
-[string[]]$InstallAptGet = @(
-    if ($env:InstallAptGet) { $env:InstallAptGet -split ',' }    
-),
+[string[]]$InstallAptGet = @($env:InstallAptGet -split ','),
 # The modules to be installed.
-[string[]]$InstallModule = @(
-    if ($env:InstallModule) { $env:InstallModule -split ',' }    
-    else {  }
-)
-)
+[string[]]$InstallModule = @($env:InstallModule -split ','),
+# The Ruby gems to be installed.
+[string[]]$InstallRubyGem = @($env:InstallRubyGem -split ','),
 
-# Get the root module directory
-$rootModuleDirectory = @($env:PSModulePath -split '[;:]')[0]
-
-# Determine the path to the module destination. 
-$moduleDestination = "$rootModuleDirectory/$ModuleName"
-# Copy the module to the destination
-# (this is being used instead of the COPY statement in Docker, to avoid additional layers).
-Copy-Item -Path "$psScriptRoot" -Destination $moduleDestination -Recurse -Force
+# If set, will keep the .git directories.
+[switch]$KeepGit = $($env:KeepGit -match $true)
+)
 
 # Copy all container-related scripts to the root of the container.
 Get-ChildItem -Path $PSScriptRoot | 
     Where-Object Name -Match '^Container\..+?\.ps1$' | 
     Copy-Item -Destination /
+
+# Create a profile
+New-Item -Path $Profile -ItemType File | Out-Null
+
+if ($ModuleName) {
+    # Get the root module directory
+    $rootModuleDirectory = @($env:PSModulePath -split '[;:]')[0]
+
+    # Determine the path to the module destination. 
+    $moduleDestination = "$rootModuleDirectory/$ModuleName"
+    # Copy the module to the destination
+    # (this is being used instead of the COPY statement in Docker, to avoid additional layers).
+    Copy-Item -Path "$psScriptRoot" -Destination $moduleDestination -Recurse -Force
+
+    # and import this module in the profile
+    Add-Content -Path $profile -Value "Import-Module $ModuleName" -Force
+}
+
+# If we have modules to install
+if ($InstallModule) { 
+    # Install the modules
+    Install-Module -Name $InstallModule -Force -AcceptLicense -Scope CurrentUser
+    # and import them in the profile
+    Add-Content -Path $Profile -Value "Import-Module '$($InstallModule -join "','")'" -Force
+}
 
 # If we have packages to install
 if ($InstallAptGet) {
@@ -70,16 +85,9 @@ if ($InstallAptGet) {
                 Out-Host
 }
 
-# Create a new profile
-New-Item -Path $Profile -ItemType File -Force |
-    # and import this module in the profile
-    Add-Content -Value "Import-Module $ModuleName" -Force
-# If we have modules to install
-if ($InstallModule) { 
-    # Install the modules
-    Install-Module -Name $InstallModule -Force -AcceptLicense -Scope CurrentUser 
-    # and import them in the profile
-    Add-Content -Path $Profile -Value "Import-Module '$($InstallModule -join "','")'" -Force
+if ($InstallRubyGem) {
+    # Install the Ruby gems
+    gem install @InstallRubyGem
 }
 
 if ($ModuleName) {
@@ -87,15 +95,17 @@ if ($ModuleName) {
     Add-Content -Path $Profile -Value "Get-Module $ModuleName | Split-Path | Push-Location" -Force
 }
 
-# Remove the .git directories from any modules
-Get-ChildItem -Path $rootModuleDirectory -Directory -Force -Recurse |
-    Where-Object Name -eq '.git' |
-    Remove-Item -Recurse -Force
+if (-not $KeepGit) {
+    # Remove the .git directories from any modules
+    Get-ChildItem -Path $rootModuleDirectory -Directory -Force -Recurse |
+        Where-Object Name -eq '.git' |
+        Remove-Item -Recurse -Force
+}
 
 # Congratulations! You have successfully initialized the container image.
 # This script should work in about any module, with minor adjustments.
 # If you have any adjustments, please put them below here, in the `#region Custom`
 
 #region Custom
-gem install jekyll
+bundle config --global silence_root_warning true
 #endregion Custom
